+++ client/src/components/pay/SuccessPayment.jsx
@@ -3,7 +3,7 @@ import { useEffect } from "react";
 import { useLocation, useNavigate, useParams } from "react-router-dom";
 import styled from "styled-components";
 
-import BasicButton from "../BasicButton";
+import BasicButton from "../button/BasicButton";
 
 const Container = styled.div`
   max-width: 1050px;
