+++ client/src/pages/cart/index.js
@@ -8,8 +8,8 @@ import { setInfo } from "../../app/reducer/productId2Pay";
 import { ReactComponent as CartIcon } from "../../assets/cart_icon.svg";
 import CartProductItem from "../../components/cart/CartProductItem";
 import CheckBox from "../../components/CheckBox";
-import BasicButton from "../../components/BasicButton";
-import ColorButton from "../../components/ColorButton";
+import BasicButton from "../../components/button/BasicButton";
+import ColorButton from "../../components/button/ColorButton";
 import ProductItemSlider from "../../components/ProductItemSlider";
 import { Title, TodayRecommendProducts } from "..";
 import noCartImg from "../../assets/no_carts.png";
