+++ client/src/pages/mypage/deliverySet.js
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import axios from "axios";
 import styled from "styled-components";
-import BasicButton from "../../components/BasicButton";
+import BasicButton from "../../components/button/BasicButton";
 import CheckBox from "../../components/CheckBox";
 import Mypagehead from "../../components/MypageHead";
 import ListLayout from "../../components/ListLayout";
