+++ client/src/pages/productDetail/productContent/Inquiry.jsx
@@ -1,7 +1,7 @@
 import { useState, useEffect } from "react";
 import { useParams } from "react-router-dom";
 import styled from "styled-components";
-import BasicButton from "../../../components/BasicButton";
+import BasicButton from "../../../components/button/BasicButton";
 import axios from "axios";
 
 import InquiryDetail from "./InquiryDetail";
